services/users/requirements.txt
Flask==1.1.1
Flask-Admin==1.5.5
Flask-RESTX==0.1.1
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
psycopg2-binary==2.8.4
werkzeug==0.16.1
flask-cors==3.0.8
flask-bcrypt==0.7.1
pyjwt==1.7.1
marshmallow==3.6.1
gspread==3.6.0
bs4==0.0.1
oauth2client==4.1.3
boto3==1.14.15
